refactor(config): move CORS related config into own file
